Marketplace/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
flask>=3.0.0
requests>=2.31.0
spyne>=2.14.0
zeep>=4.2.1
lxml>=5.1.0
